% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateGeography.R
\name{aggregateGeography}
\alias{aggregateGeography}
\title{Aggregate Geography to Higher Census Level}
\usage{
aggregateGeography(dataframe, level)
}
\arguments{
\item{dataframe}{This is the dataframe that contains your data. Each entry should be already be in census geography, and should contain the columns 'block', 'concentration' and 'area'. Columns must be named as such.}

\item{level}{This defines how high you would like to aggregate your data. Options are: 'state', 'county', 'tract', or 'block group'.}
}
\description{
This function allows you to aggregate the RSEI toxicity data from block level to any higher level. The function can only aggregate up to coarser geographies, and will
}
\examples{
aggregateGeography(data, 'state')

}
