% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrapolateMean.R
\name{extrapolateMean}
\alias{extrapolateMean}
\title{Create non-parametric estimate of what distributions would have been, given percentile placement stayed the same}
\usage{
extrapolateMean(year, old, col, log = NA, n = 10000)
}
\arguments{
\item{year}{This is the dataframe that contains your data. Each entry should be already be in census geography, and should contain the column 'concentration'.}

\item{old}{This is the dataframe for the original year you'd like to extrapolate from. Must also contain "concentration".}

\item{col}{This is a string with the name of the column that you'd like to extrapolate. Must have same name in both df's}

\item{log}{Optional, parameter to specify that you'd like to simulate log toxicity. Set to any string for log.}

\item{n}{Optional, default set to 10000, that specifies how many times you draw from the toxicities.}
}
\description{
This function allows you to estimate the mean toxicity of a group, given their placement in the overall distribution had stayed the same over time.
}
\examples{
extrapolateMean(race_data[race_data$year == year, ], race_data[race_data$year == 1990, ], "black", n = 10)
}
