% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonParametricExtrapolate.R
\name{nonParametricExtrapolate}
\alias{nonParametricExtrapolate}
\title{Create non-parametric estimate of what distributions would have been, given percentile placement stayed the same}
\usage{
nonParametricExtrapolate(year, old)
}
\arguments{
\item{old}{This is the dataframe for the original year you'd like to extrapolate from.}

\item{dataframe}{This is the dataframe that contains your data. Each entry should be already be in census geography, and should contain the column 'block', 'black' and 'white'. Columns must be named as such.}
}
\description{
This function allows you to get an estimate of the toxicity distribution for a given group, given their placement in the overall distribution had stayed the same over time.
}
\examples{
nonParametricExtrapolate(t2010, t1990)

}
