% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentilePlot.R
\name{percentilePlot}
\alias{percentilePlot}
\title{Create a plot of the cdf's of the log toxicity of various groups.}
\usage{
percentilePlot(data, value, log = "l")
}
\arguments{
\item{data}{is a list of dataframes. The list must be at least 1 long. Each of the dataframes must contain a column "concentration", and 
will be the data for various groups at different times.}

\item{value}{is a list of the columns that you would like plotted. All dataframes must contain all values included here. For example, 
if you want to plot the hispanic distributions in 1990 and 2000 -> percentilePlot(c(t1990, t2000), c("hispanic")). Comparitively, if 
you want to plot hispanic in 1990, and asian in 2000, percentilePlot(c(t1990, t2000), c("hispanic", "asian")) will plot both groups in 
both years, and throw an error if any of those columns don't exist.}

\item{log}{if log == "l", then it will plot the logged toxicity, if log == "", it will plot normal toxicity}
}
\description{
Create a plot of the cdf's of the log toxicity of various groups.
}
\examples{
percentilePlot(list(toxic_1990), c("hispanic", "white"))
}
