% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeChangePlot.R
\name{timeChangePlot}
\alias{timeChangePlot}
\title{Create a plot showing the improvement of groups relataive to the total population}
\usage{
timeChangePlot(data1, data2, values, type, grain = 20)
}
\arguments{
\item{data1}{is a full dataframe with data on toxicity concentration and population counts. The dataframe must contain the variables: 'lconcentration' and the count of total population 'tot_pop'}

\item{data2}{is the same format of data, but at a second time period, over which you are interested in the changes}

\item{values}{is a vector of strings refering to the columns you would like to see the relative changes in. All strings must match to a column in the dataframe.}

\item{type}{should be either "rank" or "value".}

\item{grain}{defines how fine grain you want the calulation to be. No input will default to 100 calculations}
}
\value{
Returns a plot object, showing the differences in changes in the groups, that can be saved or directly displayed. Returns a positive 'value' difference if group 1 improved more than group 2 
over the interest period.Returns a negative 'rank' difference if group 1 improved more than group 2 in the defined time period.
}
\description{
Create a plot showing the improvement of groups relataive to the total population
}
\examples{
timeChangePlot(t1990, t2010, c("hispanic", "white", "black"), "value", 30)
}
