% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeDifference.R
\name{timeDifference}
\alias{timeDifference}
\title{Find differences in group improvement over time}
\usage{
timeDifference(value, group1.1, group1.2, group2.1, group2.2, type)
}
\arguments{
\item{value}{is the level of toxicity you would like to find the rank difference at, or the rank you would like to find the toxicity difference at.}

\item{group1.1}{is meant to be a dataframe with 2 columns. The columns should be the toxicity in a location and the count of the population of interest in 
that location. Column names are not important, but columns must be ordered with concentration in the first column and population counts in the second column.
1.1 should be the first group at the first time.}

\item{group1.2}{should be the first group at the second time period.}

\item{group2.1}{should be the second group at the first time period.}

\item{group2.2}{should be the second group at the first time period.}

\item{type}{should be either "rank" or "value", and defines the type of difference calculation being done.}
}
\value{
Returns the difference of the changes between the two groups. Returns a positive value difference if group 1 improved more than group 2 over the interest period.
Returns a negative rank difference if group 1 improved more than group 2 in the defined time period.
}
\description{
Find differences in group improvement over time
}
\examples{
timeDifference(0.5, t1990[, c("lconc", "hispanic")], t2010[, c("lconc", "hispanic")], t1990[, c("lconc", "white")], t2010[c("lconc", "white")], "value") 
}
