% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valueDifference.R
\name{valueDifference}
\alias{valueDifference}
\title{Find value improvement for one group}
\usage{
valueDifference(group1, group2, value)
}
\arguments{
\item{group1}{is meant to be a dataframe with 2 columns. The columns should be the toxicity in a location and the count of the population of interest in 
that location. Column names are not important, but columns must be ordered with concentration in the first column and population counts in the second column.}

\item{group2}{should have the same format as group1}

\item{value}{is the percentile you want to find the toxicity difference in.}
}
\description{
Based on the paper 'Divergent Paths: Structural Change, Economic Rank, and the Evolution of Black-White Earnings Differences, 1940-2014', this function finds the 
difference in toxicity level for the same percentile in different populations. This is calculated by individually finding the toxicity in each population and 
returning the difference. The return is group1 - group 2, and users should be aware that the calls valueDifference(group1, group2, 500) and 
valueDifference(group2, group1, 500) will not return the same answer.
}
\examples{
valueDifference(group1, group2, .6)
valueDifference(data1990[, c(2, 3)], data1990[, c(2, 7)], .99)
}
